v0.2.7 - Some changes to the BitmapFont including setting the glyph data - Fonts now use floating point to calculate the advance of the next character - GuiWidget now accepts labda as an event callback - Added Cubemap texture type - getPixels and setPixels will bind the texture automatically - Added a framebuffered version of the GuiWindowOpenGL - Fixed an issue when mouse position wouldn't work correctly on repositioned GuiWindow - Added new example of rendering a skybox using cubemap texture

+++ include/ffw/gui/guifontopengl.h
@@ -22,7 +22,7 @@ namespace ffw {
 			return TrueTypeFont::getStringSize(str, lineHeight);
 		}
 		inline int getCharAdvance(wchar_t c) const override {
-			return TrueTypeFont::getChar(c).advance;
+			return (int)TrueTypeFont::getChar(c).advance;
 		}
 		inline int getSizeInPixels() const override {
 			return TrueTypeFont::getSizePixels();
